HOME BEF Changes.

+++ mxprj/System/Inc/Delay.h
@@ -2,6 +2,6 @@
 #define _DELAY_H_
 #include "GeneralDef.h"
 void setState(bool state);
-void HAL_Delay_us(uint16_t us);
+void HAL_Delay_us(uint32_t us);
 void Delay_Init();
 #endif // !_DELAY_H_
